chore: remove typescript index.from_src
Fixes a bug in ts_project where it couldn't accept generated tsconfig.json files

This continues the cleanup of the from_src variants of the public entry points.
Some notes on this one since it's tricky:
- we now use ts_project to pre-compile the .ts files that we check in as .js. It uses a generated tsconfig for funsies
- because ts_project labels the .js outputs, we can't refer to the foo.js output as well as a foo.js input in the source tree - the golden_file_test gets confused about the labels
  so we have to rename one of the outputs to index.js
