+++ internal/npm_install/test/package/BUILD.bazel
@@ -1,5 +1,5 @@
 # Generated file from yarn_install/npm_install rule.
-# See $(bazel info output_base)/external/build_bazel_rules_nodejs/internal/npm_install/generate_build_file.js
+# See rules_nodejs/internal/npm_install/generate_build_file.ts
 
 load("@build_bazel_rules_nodejs//internal/npm_install:node_module_library.bzl", "node_module_library")
