+++ packages/protractor/test/protractor-2/BUILD.bazel
@@ -2,8 +2,7 @@ load("@npm//http-server:index.bzl", "http_server")
 load("@npm_bazel_protractor//:index.from_src.bzl", "protractor_web_test_suite")
 load("@npm_bazel_rollup//:index.bzl", "rollup_bundle")
 load("@npm_bazel_terser//:index.bzl", "terser_minified")
-load("@npm_bazel_typescript//:index.bzl", "ts_config", "ts_library")
-load("@npm_bazel_typescript//:index.from_src.bzl", "ts_devserver")
+load("@npm_bazel_typescript//:index.bzl", "ts_config", "ts_devserver", "ts_library")
 
 ts_config(
     name = "tsconfig-test",
@@ -30,6 +29,8 @@ terser_minified(
 
 ts_devserver(
     name = "devserver",
+    devserver = "@build_bazel_rules_typescript//devserver:devserver_bin",
+    devserver_host = "@build_bazel_rules_typescript//devserver:devserver_bin",
     serving_path = "/bundle.min.js",
     static_files = ["index.html"],
     deps = [":app"],
